+++ NEWS.md
@@ -1,3 +1,12 @@
+# NOTATES development
+
+## Major Changes
+- Changed germline variant filtering (for reporting) to (1) NOT "benign"/"likely benign" in ClinVar (2) MAF < 1% (in gnomAD, ExAC, TGP and ESP) and (3) Non-syn. impact and (4) NOT in FLAGs
+
+## Minor changes and bug fixes
+
+***
+
 # NOTATES v1.5.0
 
 ## Major Changes